fix(editor): move clone agent to editor
